+++ packages/metagame/src/extensions/callback/callback.cairo
@@ -30,12 +30,12 @@ pub mod MetagameCallbackComponent {
         /// Called when a token's score is updated.
         /// @param token_id The token ID (packed u256)
         /// @param score The new score value
-        fn on_score_update(ref self: TContractState, token_id: u256, score: u32);
+        fn on_score_update(ref self: TContractState, token_id: u256, score: u64);
 
         /// Called when a game ends (game_over transitions to true).
         /// @param token_id The token ID (packed u256)
         /// @param final_score The final score when game ended
-        fn on_game_over(ref self: TContractState, token_id: u256, final_score: u32);
+        fn on_game_over(ref self: TContractState, token_id: u256, final_score: u64);
 
         /// Called when all objectives are completed.
         /// @param token_id The token ID (packed u256)
@@ -55,13 +55,13 @@ pub mod MetagameCallbackComponent {
         +Drop<TContractState>,
         +MetagameCallbackHooksTrait<TContractState>,
     > of IMetagameCallback<ComponentState<TContractState>> {
-        fn on_score_update(ref self: ComponentState<TContractState>, token_id: u256, score: u32) {
+        fn on_score_update(ref self: ComponentState<TContractState>, token_id: u256, score: u64) {
             let mut contract = self.get_contract_mut();
             MetagameCallbackHooksTrait::on_score_update(ref contract, token_id, score);
         }
 
         fn on_game_over(
-            ref self: ComponentState<TContractState>, token_id: u256, final_score: u32,
+            ref self: ComponentState<TContractState>, token_id: u256, final_score: u64,
         ) {
             let mut contract = self.get_contract_mut();
             MetagameCallbackHooksTrait::on_game_over(ref contract, token_id, final_score);
@@ -103,12 +103,12 @@ pub impl MetagameCallbackHooksEmptyImpl<
     TContractState,
 > of MetagameCallbackComponent::MetagameCallbackHooksTrait<TContractState> {
     fn on_score_update(
-        ref self: TContractState, token_id: u256, score: u32,
+        ref self: TContractState, token_id: u256, score: u64,
     ) { // No-op: contracts can override for custom score handling
     }
 
     fn on_game_over(
-        ref self: TContractState, token_id: u256, final_score: u32,
+        ref self: TContractState, token_id: u256, final_score: u64,
     ) { // No-op: contracts can override for custom game over handling
     }
